+++ MatrixSDK/Data/MXRoom.m
@@ -33,16 +33,11 @@ NSString *const kMXRoomDidUpdateUnreadNotification = @"kMXRoomDidUpdateUnreadNot
 @implementation MXRoom
 @synthesize mxSession;
 
-- (id)initWithRoomId:(NSString *)roomId andMatrixSession:(MXSession *)mxSession2
+- (instancetype)init
 {
     self = [super init];
     if (self)
     {
-        _roomId = roomId;
-        mxSession = mxSession2;
-
-        _liveTimeline = [[MXEventTimeline alloc] initWithRoom:self andInitialEventId:nil];
-
         _accountData = [[MXRoomAccountData alloc] init];
 
         _typingUsers = [NSArray array];
@@ -76,6 +71,12 @@ NSString *const kMXRoomDidUpdateUnreadNotification = @"kMXRoomDidUpdateUnreadNot
     return self;
 }
 
+- (id)initWithRoomId:(NSString *)roomId andMatrixSession:(MXSession *)mxSession2
+{
+    // Let's the live MXEventTimeline use its default store.
+    return [self initWithRoomId:roomId matrixSession:mxSession2 andStore:nil];
+}
+
 - (id)initWithRoomId:(NSString *)roomId andMatrixSession:(MXSession *)mxSession2 andStateEvents:(NSArray *)stateEvents andAccountData:(MXRoomAccountData*)accountData
 {
     self = [self initWithRoomId:roomId andMatrixSession:mxSession2];
@@ -88,6 +89,27 @@ NSString *const kMXRoomDidUpdateUnreadNotification = @"kMXRoomDidUpdateUnreadNot
             // Report the provided accountData.
             // Allocate a new instance if none, in order to handle room tag events for this room.
             _accountData = accountData ? accountData : [[MXRoomAccountData alloc] init];
+        }
+    }
+    return self;
+}
+
+- (id)initWithRoomId:(NSString *)roomId matrixSession:(MXSession *)mxSession2 andStore:(id<MXStore>)store
+{
+    self = [self init];
+    if (self)
+    {
+        _roomId = roomId;
+        mxSession = mxSession2;
+
+        if (store)
+        {
+            _liveTimeline = [[MXEventTimeline alloc] initWithRoom:self initialEventId:nil andStore:store];
+        }
+        else
+        {
+            // Let the timeline use the session store
+            _liveTimeline = [[MXEventTimeline alloc] initWithRoom:self andInitialEventId:nil];
         }
     }
     return self;
@@ -243,6 +265,40 @@ NSString *const kMXRoomDidUpdateUnreadNotification = @"kMXRoomDidUpdateUnreadNot
                     failure:(void (^)(NSError *error))failure
 {
     return [mxSession.matrixRestClient setRoomName:self.state.roomId name:name success:success failure:failure];
+}
+
+- (MXHTTPOperation *)setHistoryVisibility:(MXRoomHistoryVisibility)historyVisibility
+                                  success:(void (^)())success
+                                  failure:(void (^)(NSError *))failure
+{
+    return [mxSession.matrixRestClient setRoomHistoryVisibility:self.state.roomId historyVisibility:historyVisibility success:success failure:failure];
+}
+
+- (MXHTTPOperation*)setJoinRule:(MXRoomJoinRule)joinRule
+                        success:(void (^)())success
+                        failure:(void (^)(NSError *error))failure
+{
+    return [mxSession.matrixRestClient setRoomJoinRule:self.state.roomId joinRule:joinRule success:success failure:failure];
+}
+
+- (MXHTTPOperation*)setGuestAccess:(MXRoomGuestAccess)guestAccess
+                           success:(void (^)())success
+                           failure:(void (^)(NSError *error))failure
+{
+    return [mxSession.matrixRestClient setRoomGuestAccess:self.state.roomId guestAccess:guestAccess success:success failure:failure];
+}
+
+- (MXHTTPOperation*)setDirectoryVisibility:(MXRoomDirectoryVisibility)directoryVisibility
+                                   success:(void (^)())success
+                                   failure:(void (^)(NSError *error))failure
+{
+    return [mxSession.matrixRestClient setRoomDirectoryVisibility:self.state.roomId directoryVisibility:directoryVisibility success:success failure:failure];
+}
+
+- (MXHTTPOperation*)directoryVisibility:(void (^)(MXRoomDirectoryVisibility directoryVisibility))success
+                                failure:(void (^)(NSError *error))failure
+{
+    return [mxSession.matrixRestClient directoryVisibilityOfRoom:self.state.roomId success:success failure:failure];
 }
 
 - (MXHTTPOperation*)join:(void (^)())success