+++ MatrixSDK/Data/MXRoomState.h
@@ -21,6 +21,7 @@
 #import "MXRoomMember.h"
 #import "MXRoomThirdPartyInvite.h"
 #import "MXRoomPowerLevels.h"
+#import "MXEnumConstants.h"
 
 @class MXSession;
 
@@ -66,11 +67,6 @@ A copy of the list of third party invites (actually MXRoomThirdPartyInvite insta
 @property (nonatomic, readonly) MXRoomPowerLevels *powerLevels;
 
 /**
- The visibility of the room: public or, else, private
- */
-@property (nonatomic) BOOL isPublic;
-
-/**
  The aliases of this room.
  */
 @property (nonatomic, readonly) NSArray *aliases;
@@ -94,6 +90,26 @@ A copy of the list of third party invites (actually MXRoomThirdPartyInvite insta
  The avatar url of the room.
  */
 @property (nonatomic, readonly) NSString *avatar;
+
+/**
+ The history visibility of the room.
+ */
+@property (nonatomic, readonly) MXRoomHistoryVisibility historyVisibility;
+
+/**
+ The join rule of the room.
+ */
+@property (nonatomic, readonly) MXRoomJoinRule joinRule;
+
+/**
+ Shortcut to check if the self.joinRule is public.
+ */
+@property (nonatomic, readonly) BOOL isJoinRulePublic;
+
+/**
+ The guest access of the room.
+ */
+@property (nonatomic, readonly) MXRoomGuestAccess guestAccess;
 
 /**
  The display name of the room.