+++ MatrixSDK/Data/Store/MXFileStore/MXFileStore.m
@@ -20,7 +20,7 @@
 
 #import "MXFileStoreMetaData.h"
 
-NSUInteger const kMXFileVersion = 26;
+NSUInteger const kMXFileVersion = 28;
 
 NSString *const kMXFileStoreFolder = @"MXFileStore";
 NSString *const kMXFileStoreMedaDataFile = @"MXFileStore";