+++ MatrixSDK/Utils/MXTools.m
@@ -35,6 +35,8 @@
                  kMXEventTypeStringRoomPowerLevels: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomPowerLevels],
                  kMXEventTypeStringRoomAliases: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomAliases],
                  kMXEventTypeStringRoomCanonicalAlias: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomCanonicalAlias],
+                 kMXEventTypeStringRoomHistoryVisibility: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomHistoryVisibility],
+                 kMXEventTypeStringRoomGuestAccess: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomGuestAccess],
                  kMXEventTypeStringRoomMessage: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomMessage],
                  kMXEventTypeStringRoomMessageFeedback: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomMessageFeedback],
                  kMXEventTypeStringRoomRedaction: [NSNumber numberWithUnsignedInteger:MXEventTypeRoomRedaction],
@@ -149,6 +151,27 @@
 + (BOOL)isEmailAddress:(NSString *)inputString
 {
     NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,}$" options:NSRegularExpressionCaseInsensitive error:nil];
+    
+    return (nil != [regex firstMatchInString:inputString options:0 range:NSMakeRange(0, inputString.length)]);
+}
+
++ (BOOL)isMatrixUserIdentifier:(NSString *)inputString
+{
+    NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"^@[A-Z0-9]+:[A-Z0-9.-]+\\.[A-Z]{2,}$" options:NSRegularExpressionCaseInsensitive error:nil];
+    
+    return (nil != [regex firstMatchInString:inputString options:0 range:NSMakeRange(0, inputString.length)]);
+}
+
++ (BOOL)isMatrixRoomAlias:(NSString *)inputString
+{
+    NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"^#[A-Z0-9._%+-]+:[A-Z0-9.-]+\\.[A-Z]{2,}$" options:NSRegularExpressionCaseInsensitive error:nil];
+    
+    return (nil != [regex firstMatchInString:inputString options:0 range:NSMakeRange(0, inputString.length)]);
+}
+
++ (BOOL)isMatrixRoomIdentifier:(NSString *)inputString
+{
+    NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"^![A-Z0-9]+:[A-Z0-9.-]+\\.[A-Z]{2,}$" options:NSRegularExpressionCaseInsensitive error:nil];
     
     return (nil != [regex firstMatchInString:inputString options:0 range:NSMakeRange(0, inputString.length)]);
 }